Fix yet more warnings from gcc 4.6:
block/block_compact.c:193:165: error: conversion to 'long unsigned int' from 'off_t' may change the sign of the result [-Werror=sign-conversion]
block/block_compact.c:203:170: error: conversion to 'long unsigned int' from 'off_t' may change the sign of the result [-Werror=sign-conversion]

+++ src/block/block_compact.c
@@ -194,7 +194,7 @@ __block_dump_avail(WT_SESSION_IMPL *session, WT_BLOCK *block)
 		    session, "%2u%%: %12" PRIuMAX "MB, (%" PRIuMAX "B, %"
 		    PRIuMAX "%%)",
 		    i, (uintmax_t)v / WT_MEGABYTE, (uintmax_t)v,
-		    (uintmax_t)((v * 100) / el->bytes)));
+		    (uintmax_t)((v * 100) / (off_t)el->bytes)));
 	}
 #ifdef __VERBOSE_OUTPUT_PERCENTILE
 #endif
@@ -204,7 +204,7 @@ __block_dump_avail(WT_SESSION_IMPL *session, WT_BLOCK *block)
 		    session, "%2u%%: %12" PRIuMAX "MB, (%" PRIuMAX "B, %"
 		    PRIuMAX "%%)",
 		    i * 10, (uintmax_t)v / WT_MEGABYTE, (uintmax_t)v,
-		    (uintmax_t)((v * 100) / el->bytes)));
+		    (uintmax_t)((v * 100) / (off_t)el->bytes)));
 	}
 
 	return (0);
